+++ finsync-user-service/src/main/java/in/syncuser/entity/User.java
@@ -28,7 +28,7 @@ public class User extends BaseEntity{
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String firstName;
 	private String lastName;
 	private String gender;
@@ -73,7 +73,7 @@ public class User extends BaseEntity{
 		this.fullName = this.firstName +" "+this.lastName;
 	}
 
-	public User(Long id) {
+	public User(Integer id) {
 		super();
 		this.id = id;
 	}
