+++ finsync-user-service/src/main/java/in/syncuser/service/RoleService.java
@@ -13,7 +13,7 @@ public interface RoleService {
 
 	List<GrantedAuthority> fetchAll();
 
-	List<RoleApiDTO> fetchByUID(Long id);
+	List<RoleApiDTO> fetchByUID(Integer id);
 
 	GrantedAuthority insertRole(RoleApiDTO apiModel);
